Split buildpacks-jvm-shared into modules
