+++ buildpacks/sbt/src/errors.rs
@@ -2,7 +2,7 @@ use crate::configuration::ReadSbtBuildpackConfigurationError;
 use crate::layers::sbt_extras::SbtExtrasLayerError;
 use crate::layers::sbt_global::SbtGlobalLayerError;
 use crate::sbt_version::ReadSbtVersionError;
-use buildpacks_jvm_shared::log_please_try_again_error;
+use buildpacks_jvm_shared::log::log_please_try_again_error;
 use buildpacks_jvm_shared::system_properties::ReadSystemPropertiesError;
 use indoc::formatdoc;
 use libherokubuildpack::log::log_error;
