+++ shared/src/system_properties.rs
@@ -1,4 +1,4 @@
-use crate::none_on_not_found;
+use crate::result::none_on_not_found;
 use std::collections::HashMap;
 use std::fs;
 use std::path::Path;
